[Simpyfy] neo4j node parser util로 분리, feed에서 사용하지 않는 쿼리 삭제 #24
why
결과 파싱하는 부분은 반복적으로 사용될 것임, 사용되지 않는 쿼리로 확인할 때 헷갈림

how
util 파일로 분리하고 사용자지 않는 쿼리 삭제
